refactor: move agent runtime payloads to protocols
