+++ backend/protocols/__init__.py
@@ -1,0 +1,1 @@
+"""Transport protocol contracts shared across backend boundaries."""
