+++ backend/web/routers/threads.py
@@ -1036,8 +1036,9 @@ async def send_message(
     if not payload.message.strip():
         raise HTTPException(status_code=400, detail="message cannot be empty")
 
+    from backend.protocols.agent_runtime import AgentThreadInputEnvelope
     from backend.web.services.agent_pool import get_or_create_agent, resolve_thread_sandbox
-    from backend.web.services.agent_runtime_gateway import AgentThreadInputEnvelope, NativeAgentRuntimeGateway
+    from backend.web.services.agent_runtime_gateway import NativeAgentRuntimeGateway
 
     message = payload.message
     # @@@attachment-wire - sync files to sandbox and prepend paths
@@ -1172,7 +1173,8 @@ async def resolve_thread_permission_request(
 
     followup: dict[str, Any] | None = None
     if is_ask_user_question and payload.decision == "allow" and pending_request is not None and answers is not None:
-        from backend.web.services.agent_runtime_gateway import AgentThreadInputEnvelope, NativeAgentRuntimeGateway
+        from backend.protocols.agent_runtime import AgentThreadInputEnvelope
+        from backend.web.services.agent_runtime_gateway import NativeAgentRuntimeGateway
 
         answered_payload = _build_ask_user_question_answered_payload(
             pending_request,
