+++ tests/Unit/backend/web/services/test_agent_runtime_gateway_chat_delivery.py
@@ -5,14 +5,14 @@ from typing import Any
 
 import pytest
 
-from backend.web.services.agent_runtime_gateway import (
+from backend.protocols.agent_runtime import (
     AgentChatActor,
     AgentChatContext,
     AgentChatDeliveryEnvelope,
     AgentChatMessage,
     AgentChatRecipient,
-    NativeAgentRuntimeGateway,
 )
+from backend.web.services.agent_runtime_gateway import NativeAgentRuntimeGateway
 from core.runtime.middleware.monitor import AgentState
